feat: hoàn thiện chỉnh sửa giao diện màn notification
