+++ app/src/main/java/com/pwhs/quickmem/presentation/app/notification/NotificationUiAction.kt
@@ -1,7 +1,8 @@
 package com.pwhs.quickmem.presentation.app.notification
 
 sealed class NotificationUiAction {
-    data object LoadNotifications : NotificationUiAction()
+    data class SetArgs(val userId: String) : NotificationUiAction()
+    data class LoadNotifications(val userId: String) : NotificationUiAction()
     data class MarkAsRead(val notificationId: String) : NotificationUiAction()
     data class DeleteNotification(val notificationId: String) : NotificationUiAction()
     data object RefreshNotifications : NotificationUiAction()
